+++ ModelImporter/import_scene.py
@@ -7,6 +7,7 @@ import subprocess
 
 # Blender imports
 import bpy  # pylint: disable=import-error
+import bmesh  # pylint: disable=import-error
 from mathutils import Matrix, Vector  # pylint: disable=import-error
 
 # Internal imports
@@ -126,6 +127,29 @@ class ImportScene():
 
         # get the information about what data the geometry file contains
         with open(self.geometry_file, 'rb') as f:
+            # Check to see if we have any mesh collision data
+            f.seek(0x6C)
+            self.CollisionIndexCount = struct.unpack('<I', f.read(0x4))[0]
+            if self.CollisionIndexCount != 0:
+                # Determine if the index data is 16bit or 32 bit (2 or 4 bytes)
+                f.seek(0x68)
+                self.Indices16Bit = bool(struct.unpack('<I', f.read(0x4))[0])
+                f.seek(0x180)
+                list_offset, _ = read_list_header(f)
+                f.seek(list_offset, 1)
+                if self.Indices16Bit:
+                    fmt = 'H'
+                    mult = 2
+                else:
+                    fmt = 'I'
+                    mult = 4
+                # Read all the mesh index data into a single list
+                self.mesh_indexes = struct.unpack(
+                    '<' + fmt * self.CollisionIndexCount,
+                    f.read(self.CollisionIndexCount * mult))
+            else:
+                self.mesh_indexes = list()
+
             f.seek(0x140)
             self.count, self.stride = struct.unpack('<II', f.read(0x8))
             # skip platform data
@@ -159,6 +183,19 @@ class ImportScene():
         mesh_node._generate_geometry()
         mesh_node._generate_bounded_hull(self.bh_data)
 
+    def load_collision_mesh(self, mesh_node):
+        """ Load the collision mesh data.
+        This only needs the bounded hull data and the index buffer with the
+        VERTRSTART vaue subtracted off.
+        """
+        idx_start = int(mesh_node.Attribute('BATCHSTART'))
+        idx_count = int(mesh_node.Attribute('BATCHCOUNT'))
+        idxs = self.mesh_indexes[idx_start: idx_start + idx_count]
+        for idx in idxs:
+            mesh_node.idxs.append(idx - int(mesh_node.Attribute('VERTRSTART')))
+        mesh_node._generate_geometry(from_bh=True)
+        mesh_node._generate_bounded_hull(self.bh_data)
+
     def render_mesh(self, mesh_ID):
         """Render the specified mesh in the blender view. """
         obj = self.scene_node_data.get(mesh_ID)
@@ -187,6 +224,13 @@ class ImportScene():
             elif (obj.Type == 'LOCATOR' or obj.Type == 'JOINT'
                   or obj.Type == 'REFERENCE'):
                 self._add_empty_to_scene(obj)
+            elif obj.Type == 'COLLISION':
+                if self.settings['import_collisions']:
+                    if obj.Attribute('TYPE') == 'Mesh':
+                        self.load_collision_mesh(obj)
+                        self._add_mesh_collision_to_scene(obj)
+                    else:
+                        self._add_primitive_collision_to_scene(obj)
         self.state = {'FINISHED'}
 
 # region private methods
@@ -216,7 +260,7 @@ class ImportScene():
             descriptor_name = op.basename(self.scene_name) + '.DESCRIPTOR.MBIN'
             if op.exists(op.join(self.local_directory, descriptor_name)):
                 empty_obj.NMSReference_props.is_proc = True
-            self.local_objects[empty_obj.name] = empty_obj
+            self.local_objects[scene_node] = empty_obj
             return
 
         # Otherwise just assign everything as usual...
@@ -238,7 +282,7 @@ class ImportScene():
         scale = scene_node.Transform['Scale']
         empty_obj.scale = Vector(scale)
 
-        self.local_objects[name] = empty_obj
+        self.local_objects[scene_node] = empty_obj
 
         if not standalone:
             if self.parent_obj is not None and scene_node.parent.Name is None:
@@ -247,7 +291,7 @@ class ImportScene():
                 self.ref_scenes[self.scene_name].append(empty_obj)
             elif scene_node.parent.Name is not None:
                 # Other child
-                empty_obj.parent = self.local_objects[scene_node.parent.Name]
+                empty_obj.parent = self.local_objects[scene_node.parent]
             else:
                 # Direct child of loaded scene
                 empty_obj.parent = self.local_objects[self.scene_basename]
@@ -273,6 +317,116 @@ class ImportScene():
             else:
                 print("The reference node {0} has a reference to a path "
                       "that doesn't exist ({1})".format(name, ref_scene_path))
+
+    def _add_mesh_collision_to_scene(self, scene_node):
+        """ Adds the given collision node to the Blender scene. """
+        name = scene_node.Name + '_COLL'
+        mesh = bpy.data.meshes.new(name)
+        mesh.from_pydata(scene_node.bounded_hull,
+                         scene_node.edges,
+                         scene_node.faces)
+        bh_obj = bpy.data.objects.new(name, mesh)
+
+        bh_obj.NMSNode_props.node_types = 'Collision'
+        bh_obj.NMSCollision_props.collision_types = 'Mesh'
+
+        # get transform and apply
+        transform = scene_node.Transform['Trans']
+        bh_obj.location = Vector(transform)
+        # get rotation and apply
+        rotation = scene_node.Transform['Rot']
+        bh_obj.rotation_mode = 'ZXY'
+        bh_obj.rotation_euler = rotation
+        # get scale and apply
+        scale = scene_node.Transform['Scale']
+        bh_obj.scale = Vector(scale)
+
+        if self.parent_obj is not None and scene_node.parent.Name is None:
+            # Direct child of reference node
+            bh_obj.parent = self.parent_obj
+            self.ref_scenes[self.scene_name].append(bh_obj)
+        elif scene_node.parent.Name is not None:
+            # Other child
+            parent_obj = self.local_objects[scene_node.parent]
+            bh_obj.parent = parent_obj
+        else:
+            # Direct child of loaded scene
+            bh_obj.parent = self.local_objects[self.scene_basename]
+
+        if not self.settings['show_collisions']:
+            # Only draw the collisions if they are wanted
+            bh_obj.hide = True
+
+        self.scn.objects.link(bh_obj)
+        self.local_objects[scene_node] = bh_obj
+
+    def _add_primitive_collision_to_scene(self, scene_node):
+        name = scene_node.Name + '_COLL'
+        mesh = bpy.data.meshes.new(name)
+        coll_type = scene_node.Attribute('TYPE')
+        bm = bmesh.new()
+        if coll_type == 'Box':
+            bmesh.ops.create_cube(bm, size=1.0)
+            scale_mult = [float(scene_node.Attribute('WIDTH')),
+                          float(scene_node.Attribute('HEIGHT')),
+                          float(scene_node.Attribute('DEPTH'))]
+        elif coll_type == 'Sphere':
+            bmesh.ops.create_icosphere(bm, subdivisions=4, diameter=1.0)
+            scale_mult = [float(scene_node.Attribute('RADIUS')),
+                          float(scene_node.Attribute('RADIUS')),
+                          float(scene_node.Attribute('RADIUS'))]
+        elif coll_type == 'Cylinder':
+            bmesh.ops.create_cone(bm, cap_ends=True, cap_tris=True,
+                                  diameter1=1.0, diameter2=1.0, depth=1.0,
+                                  segments=20, matrix=ROT_MATRIX)
+            scale_mult = [float(scene_node.Attribute('RADIUS')),
+                          float(scene_node.Attribute('HEIGHT')),
+                          float(scene_node.Attribute('RADIUS'))]
+        else:
+            print('Skipping unsupported coll_type: {0}'.format(coll_type))
+            bm.free()
+            del bm
+            return
+        # Convert the bmesh back to the mesh
+        bm.to_mesh(mesh)
+        bm.free()
+        del bm
+
+        coll_obj = bpy.data.objects.new(name, mesh)
+        coll_obj.NMSNode_props.node_types = 'Collision'
+        coll_obj.NMSCollision_props.collision_types = coll_type
+
+        # get transform and apply
+        transform = scene_node.Transform['Trans']
+        coll_obj.location = Vector(transform)
+        # get rotation and apply
+        rotation = scene_node.Transform['Rot']
+        coll_obj.rotation_mode = 'ZXY'
+        coll_obj.rotation_euler = rotation
+        # get scale and apply
+        scale = scene_node.Transform['Scale']
+        mod_scale = tuple(scale_mult[i] * scale[i] for i in range(3))
+        coll_obj.scale = Vector(mod_scale)
+
+        # TODO: make its own function
+        if self.parent_obj is not None and scene_node.parent.Name is None:
+            # Direct child of reference node
+            coll_obj.parent = self.parent_obj
+            self.ref_scenes[self.scene_name].append(coll_obj)
+        elif scene_node.parent.Name is not None:
+            # Other child
+            parent_obj = self.local_objects[scene_node.parent]
+            coll_obj.parent = parent_obj
+        else:
+            # Direct child of loaded scene
+            coll_obj.parent = self.local_objects[self.scene_basename]
+
+        if not self.settings['show_collisions']:
+            # Only draw the collisions if they are wanted
+            coll_obj.hide = True
+
+        self.scn.objects.link(coll_obj)
+        self.local_objects[scene_node] = coll_obj
 
     def _add_existing_to_scene(self):
         # existing is a list of child objects to the reference
@@ -317,7 +471,7 @@ class ImportScene():
         scale = scene_node.Transform['Scale']
         mesh_obj.scale = Vector(scale)
 
-        self.local_objects[name] = mesh_obj
+        self.local_objects[scene_node] = mesh_obj
 
         # give object correct parent
         if not standalone:
@@ -327,7 +481,7 @@ class ImportScene():
                 self.ref_scenes[self.scene_name].append(mesh_obj)
             elif scene_node.parent.Name is not None:
                 # Other child
-                parent_obj = self.local_objects[scene_node.parent.Name]
+                parent_obj = self.local_objects[scene_node.parent]
                 mesh_obj.parent = parent_obj
             else:
                 # Direct child of loaded scene
@@ -695,7 +849,11 @@ class ImportScene():
         elif size == 2:
             fmt = '<H'
         else:
-            raise ValueError('Something has gone wrong!!')
+            err = ("An error has ocurred. Here is the object information:\n" +
+                   "Mesh name: {0}\n".format(mesh.Name) +
+                   "Mesh indexes: {0}\n".format(idx_count) +
+                   "Mesh metadata: {0}\n".format(mesh.metadata))
+            raise ValueError(err)
         with open(self.geometry_stream_file, 'rb') as f:
             f.seek(mesh.metadata.idx_off)
             for _ in range(idx_count):