+++ __init__.py
@@ -1,17 +1,24 @@
 import bpy
 from .BlenderExtensions import (NMSNodes, NMSEntities, NMSPanels,
                                 NMSShaderNode, SettingsPanels)
+from bpy.props import PointerProperty  # noqa pylint: disable=import-error, no-name-in-module
 
+# External API operators
 from .NMSDK import ImportSceneOperator, ImportMeshOperator
+# Main IO operators
 from .NMSDK import NMS_Export_Operator, NMS_Import_Operator
-from .NMSDK import _FixOldFormat
+# Internal operators
+from .NMSDK import (_FixOldFormat, _ToggleCollisionVisibility,
+                    _SaveDefaultSettings)
+# Settings
+from .NMSDK import NMSDKSettings, NMSDKDefaultSettings
 
 customNodes = NMSNodes()
 
 bl_info = {
     "name": "No Man's Sky Development Kit",
     "author": "gregkwaste, monkeyman192",
-    "version": (0, 9, 11),
+    "version": (0, 9, 12),
     "blender": (2, 79, 0),
     "location": "File > Export",
     "description": "Create NMS scene structures and export to NMS File format",
@@ -35,6 +42,16 @@ def menu_func_import(self, context):
 def register():
     bpy.utils.register_class(NMS_Export_Operator)
     bpy.utils.register_class(NMS_Import_Operator)
+    bpy.utils.register_class(NMSDKSettings)
+    bpy.utils.register_class(NMSDKDefaultSettings)
+    bpy.utils.register_class(ImportSceneOperator)
+    bpy.utils.register_class(ImportMeshOperator)
+    bpy.utils.register_class(_FixOldFormat)
+    bpy.utils.register_class(_ToggleCollisionVisibility)
+    bpy.utils.register_class(_SaveDefaultSettings)
+    bpy.types.Scene.nmsdk_settings = PointerProperty(type=NMSDKSettings)
+    bpy.types.Scene.nmsdk_default_settings = PointerProperty(
+        type=NMSDKDefaultSettings)
     bpy.types.INFO_MT_file_export.append(menu_func_export)
     bpy.types.INFO_MT_file_import.append(menu_func_import)
     NMSPanels.register()
@@ -42,14 +59,20 @@ def register():
     customNodes.register()
     NMSEntities.register()
     SettingsPanels.register()
-    bpy.utils.register_class(ImportSceneOperator)
-    bpy.utils.register_class(ImportMeshOperator)
-    bpy.utils.register_class(_FixOldFormat)
 
 
 def unregister():
     bpy.utils.unregister_class(NMS_Export_Operator)
     bpy.utils.unregister_class(NMS_Import_Operator)
+    bpy.utils.unregister_class(NMSDKSettings)
+    bpy.utils.unregister_class(NMSDKDefaultSettings)
+    bpy.utils.unregister_class(ImportSceneOperator)
+    bpy.utils.unregister_class(ImportMeshOperator)
+    bpy.utils.unregister_class(_FixOldFormat)
+    bpy.utils.unregister_class(_ToggleCollisionVisibility)
+    bpy.utils.unregister_class(_SaveDefaultSettings)
+    del bpy.types.Scene.nmsdk_settings
+    del bpy.types.Scene.nmsdk_default_settings
     bpy.types.INFO_MT_file_export.remove(menu_func_export)
     bpy.types.INFO_MT_file_import.remove(menu_func_import)
     NMSPanels.unregister()
@@ -57,6 +80,3 @@ def unregister():
     customNodes.unregister()
     NMSEntities.unregister()
     SettingsPanels.unregister()
-    bpy.utils.unregister_class(ImportSceneOperator)
-    bpy.utils.unregister_class(ImportMeshOperator)
-    bpy.utils.unregister_class(_FixOldFormat)